ensure integration tests fail when defined in multiple files
